backend/requirements.txt
fastapi
uvicorn
python-multipart
aiohttp
aiosqlite
requests
python-telegram-bot==20.7
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy